Better separation between GPS and NAV
