+++ src/main/blackbox/blackbox.c
@@ -54,7 +54,6 @@
 
 #include "flight/failsafe.h"
 #include "flight/mixer.h"
-#include "flight/navigation.h"
 #include "flight/pid.h"
 #include "flight/servos.h"
