+++ src/main/flight/navigation.c
@@ -22,7 +22,7 @@
 
 #include "platform.h"
 
-#ifdef USE_GPS
+#ifdef USE_NAV
 
 #include "build/debug.h"
 
@@ -72,16 +72,11 @@ bool areSticksInApModePosition(uint16_t ap_mode);
 // **********************
 // GPS
 // **********************
-int16_t GPS_angle[ANGLE_INDEX_COUNT] = { 0, 0 };    // it's the angles that must be applied for GPS correction
-int32_t GPS_home[2];
 int32_t GPS_hold[2];
 
-uint16_t GPS_distanceToHome;        // distance to home point in meters
-int16_t GPS_directionToHome;        // direction to home or hol point in degrees
 
 static int16_t nav[2];
 static int16_t nav_rated[2];               // Adding a rate controller to the navigation to make it smoother
-navigationMode_e nav_mode = NAV_MODE_NONE;    // Navigation mode
 
 // When using PWM input GPS usage reduces number of available channels by 2 - see pwm_common.c/pwmInit()
 void navigationInit(void)
@@ -104,23 +99,17 @@ void navigationInit(void)
 #define NAV_TAIL_FIRST             0    // true - copter comes in with tail first
 #define NAV_SET_TAKEOFF_HEADING    1    // true - when copter arrives to home position it rotates it's head to takeoff direction
 
-#define GPS_FILTERING              1    // add a 5 element moving average filter to GPS coordinates, helps eliminate gps noise but adds latency
 #define GPS_LOW_SPEED_D_FILTER     1    // below .5m/s speed ignore D term for POSHOLD_RATE, theoretically this also removed D term induced noise
 
-static void GPS_distance_cm_bearing(int32_t * lat1, int32_t * lon1, int32_t * lat2, int32_t * lon2, uint32_t * dist, int32_t * bearing);
 //static void GPS_distance(int32_t lat1, int32_t lon1, int32_t lat2, int32_t lon2, uint16_t* dist, int16_t* bearing);
-static void GPS_calc_longitude_scaling(int32_t lat);
-static void GPS_calc_velocity(void);
 static void GPS_calc_location_error(int32_t * target_lat, int32_t * target_lng, int32_t * gps_lat, int32_t * gps_lng);
 
-#ifdef USE_NAV
 static bool check_missed_wp(void);
 static void GPS_calc_poshold(void);
 static void GPS_calc_nav_rate(uint16_t max_speed);
 static void GPS_update_crosstrack(void);
 static uint16_t GPS_calc_desired_speed(uint16_t max_speed, bool _slow);
 static int32_t wrap_36000(int32_t angle);
-#endif
 
 static int32_t wrap_18000(int32_t error);
 
@@ -149,7 +138,6 @@ typedef struct {
 static PID posholdPID[2];
 static PID poshold_ratePID[2];
 
-#ifdef USE_NAV
 static PID_PARAM navPID_PARAM;
 static PID navPID[2];
 
@@ -181,7 +169,6 @@ static int32_t get_D(int32_t input, float *dt, PID *pid, PID_PARAM *pid_param)
     // add in derivative component
     return pid_param->kD * pid->derivative;
 }
-#endif
 
 static void reset_PID(PID *pid)
 {
@@ -190,8 +177,6 @@ static void reset_PID(PID *pid)
     pid->last_derivative = 0;
 }
 
-#define GPS_X 1
-#define GPS_Y 0
 
 /****************** PI and PID controllers for GPS ********************///32938 -> 33160
 
@@ -200,18 +185,13 @@ static void reset_PID(PID *pid)
 #define NAV_SLOW_NAV               true
 #define NAV_BANK_MAX               3000 // 30deg max banking when navigating (just for security and testing)
 
-static float dTnav;             // Delta Time in milliseconds for navigation computations, updated with every good GPS read
-static int16_t actual_speed[2] = { 0, 0 };
-static float GPS_scaleLonDown = 1.0f;  // this is used to offset the shrinking longitude as we go towards the poles
 static int32_t error[2];
 
-#ifdef USE_NAV
 // The difference between the desired rate of travel and the actual rate of travel
 // updated after GPS read - 5-10hz
 static int16_t rate_error[2];
 // The amount of angle correction applied to target_bearing to bring the copter back on its optimum path
 static int16_t crosstrack_error;
-#endif
 
 // Currently used WP
 static int32_t GPS_WP[2];
@@ -238,95 +218,14 @@ static uint32_t wp_distance;
 // used for slow speed wind up when start navigation;
 static int16_t waypoint_speed_gov;
 
-////////////////////////////////////////////////////////////////////////////////////
-// moving average filter variables
-//
-#define GPS_FILTER_VECTOR_LENGTH 5
-
-static uint8_t GPS_filter_index = 0;
-static int32_t GPS_filter[2][GPS_FILTER_VECTOR_LENGTH];
-static int32_t GPS_filter_sum[2];
-static int32_t GPS_read[2];
-static int32_t GPS_filtered[2];
-static int32_t GPS_degree[2];   //the lat lon degree without any decimals (lat/10 000 000)
-static uint16_t fraction3[2];
-
 // This is the angle from the copter to the "next_WP" location
 // with the addition of Crosstrack error in degrees * 100
 static int32_t nav_bearing;
 // saves the bearing at takeof (1deg = 1) used to rotate to takeoff direction when arrives at home
-static int16_t nav_takeoff_bearing;
 
-void GPS_calculateDistanceAndDirectionToHome(void)
+
+void navNewGpsData(void)
 {
-    if (STATE(GPS_FIX_HOME)) {      // If we don't have home set, do not display anything
-        uint32_t dist;
-        int32_t dir;
-        GPS_distance_cm_bearing(&gpsSol.llh.lat, &gpsSol.llh.lon, &GPS_home[LAT], &GPS_home[LON], &dist, &dir);
-        GPS_distanceToHome = dist / 100;
-        GPS_directionToHome = dir / 100;
-    } else {
-        GPS_distanceToHome = 0;
-        GPS_directionToHome = 0;
-    }
-}
-
-void onGpsNewData(void)
-{
-    static uint32_t nav_loopTimer;
-
-    if (!(STATE(GPS_FIX) && gpsSol.numSat >= 5)) {
-        return;
-    }
-
-    if (!ARMING_FLAG(ARMED))
-        DISABLE_STATE(GPS_FIX_HOME);
-
-    if (!STATE(GPS_FIX_HOME) && ARMING_FLAG(ARMED))
-        GPS_reset_home_position();
-
-    // Apply moving average filter to GPS data
-#if defined(GPS_FILTERING)
-    GPS_filter_index = (GPS_filter_index + 1) % GPS_FILTER_VECTOR_LENGTH;
-    for (int axis = 0; axis < 2; axis++) {
-        GPS_read[axis] = axis == LAT ? gpsSol.llh.lat : gpsSol.llh.lon; // latest unfiltered data is in GPS_latitude and GPS_longitude
-        GPS_degree[axis] = GPS_read[axis] / 10000000;   // get the degree to assure the sum fits to the int32_t
-
-        // How close we are to a degree line ? its the first three digits from the fractions of degree
-        // later we use it to Check if we are close to a degree line, if yes, disable averaging,
-        fraction3[axis] = (GPS_read[axis] - GPS_degree[axis] * 10000000) / 10000;
-
-        GPS_filter_sum[axis] -= GPS_filter[axis][GPS_filter_index];
-        GPS_filter[axis][GPS_filter_index] = GPS_read[axis] - (GPS_degree[axis] * 10000000);
-        GPS_filter_sum[axis] += GPS_filter[axis][GPS_filter_index];
-        GPS_filtered[axis] = GPS_filter_sum[axis] / GPS_FILTER_VECTOR_LENGTH + (GPS_degree[axis] * 10000000);
-        if (nav_mode == NAV_MODE_POSHOLD) {             // we use gps averaging only in poshold mode...
-            if (fraction3[axis] > 1 && fraction3[axis] < 999) {
-                if (axis == LAT) {
-                    gpsSol.llh.lat = GPS_filtered[LAT];
-                } else {
-                    gpsSol.llh.lon = GPS_filtered[LON];
-                }
-            }
-        }
-    }
-#endif
-
-    //
-    // Calculate time delta for navigation loop, range 0-1.0f, in seconds
-    //
-    // Time for calculating x,y speed and navigation pids
-    dTnav = (float)(millis() - nav_loopTimer) / 1000.0f;
-    nav_loopTimer = millis();
-    // prevent runup from bad GPS
-    dTnav = MIN(dTnav, 1.0f);
-
-    GPS_calculateDistanceAndDirectionToHome();
-
-    // calculate the current velocity based on gps coordinates continously to get a valid speed at the moment when we start navigating
-    GPS_calc_velocity();
-
-#ifdef USE_NAV
     if (FLIGHT_MODE(GPS_HOLD_MODE) || FLIGHT_MODE(GPS_HOME_MODE)) {
         // we are navigating
 
@@ -367,19 +266,6 @@ void onGpsNewData(void)
             break;
         }
     }                   //end of gps calcs
-#endif
-}
-
-void GPS_reset_home_position(void)
-{
-    if (STATE(GPS_FIX) && gpsSol.numSat >= 5) {
-        GPS_home[LAT] = gpsSol.llh.lat;
-        GPS_home[LON] = gpsSol.llh.lon;
-        GPS_calc_longitude_scaling(gpsSol.llh.lat); // need an initial value for distance and bearing calc
-        nav_takeoff_bearing = DECIDEGREES_TO_DEGREES(attitude.values.yaw);              // save takeoff heading
-        // Set ground altitude
-        ENABLE_STATE(GPS_FIX_HOME);
-    }
 }
 
 // reset navigation (stop the navigation processor, and clear nav)
@@ -393,9 +279,7 @@ void GPS_reset_nav(void)
         nav[i] = 0;
         reset_PID(&posholdPID[i]);
         reset_PID(&poshold_ratePID[i]);
-#ifdef USE_NAV
         reset_PID(&navPID[i]);
-#endif
     }
 }
 
@@ -411,12 +295,10 @@ void gpsUsePIDs(pidProfile_t *pidProfile)
     poshold_ratePID_PARAM.kD = (float)pidProfile->pid[PID_POSR].D / 1000.0f;
     poshold_ratePID_PARAM.Imax = POSHOLD_RATE_IMAX * 100;
 
-#ifdef USE_NAV
     navPID_PARAM.kP = (float)pidProfile->pid[PID_NAVR].P / 10.0f;
     navPID_PARAM.kI = (float)pidProfile->pid[PID_NAVR].I / 100.0f;
     navPID_PARAM.kD = (float)pidProfile->pid[PID_NAVR].D / 1000.0f;
     navPID_PARAM.Imax = POSHOLD_RATE_IMAX * 100;
-#endif
 }
 
 // OK here is the onboard GPS code
@@ -431,12 +313,6 @@ void gpsUsePIDs(pidProfile_t *pidProfile)
 // this is used to offset the shrinking longitude as we go towards the poles
 // It's ok to calculate this once per waypoint setting, since it changes a little within the reach of a multicopter
 //
-static void GPS_calc_longitude_scaling(int32_t lat)
-{
-    float rads = (ABS((float)lat) / 10000000.0f) * 0.0174532925f;
-    GPS_scaleLonDown = cos_approx(rads);
-}
-
 ////////////////////////////////////////////////////////////////////////////////////
 // Sets the waypoint to navigate, reset neccessary variables and calculate initial values
 //
@@ -454,7 +330,6 @@ void GPS_set_next_wp(int32_t *lat, int32_t *lon)
     waypoint_speed_gov = navigationConfig()->nav_speed_min;
 }
 
-#ifdef USE_NAV
 ////////////////////////////////////////////////////////////////////////////////////
 // Check if we missed the destination somehow
 //
@@ -465,24 +340,7 @@ static bool check_missed_wp(void)
     temp = wrap_18000(temp);
     return (ABS(temp) > 10000); // we passed the waypoint by 100 degrees
 }
-#endif
 
-#define DISTANCE_BETWEEN_TWO_LONGITUDE_POINTS_AT_EQUATOR_IN_HUNDREDS_OF_KILOMETERS 1.113195f
-#define TAN_89_99_DEGREES 5729.57795f
-
-////////////////////////////////////////////////////////////////////////////////////
-// Get distance between two points in cm
-// Get bearing from pos1 to pos2, returns an 1deg = 100 precision
-static void GPS_distance_cm_bearing(int32_t *currentLat1, int32_t *currentLon1, int32_t *destinationLat2, int32_t *destinationLon2, uint32_t *dist, int32_t *bearing)
-{
-    float dLat = *destinationLat2 - *currentLat1; // difference of latitude in 1/10 000 000 degrees
-    float dLon = (float)(*destinationLon2 - *currentLon1) * GPS_scaleLonDown;
-    *dist = sqrtf(sq(dLat) + sq(dLon)) * DISTANCE_BETWEEN_TWO_LONGITUDE_POINTS_AT_EQUATOR_IN_HUNDREDS_OF_KILOMETERS;
-
-    *bearing = 9000.0f + atan2_approx(-dLat, dLon) * TAN_89_99_DEGREES;      // Convert the output radians to 100xdeg
-    if (*bearing < 0)
-        *bearing += 36000;
-}
 
 ////////////////////////////////////////////////////////////////////////////////////
 // keep old calculation function for compatibility (could be removed later) distance in meters, bearing in degree
@@ -494,32 +352,6 @@ static void GPS_distance_cm_bearing(int32_t *currentLat1, int32_t *currentLon1, 
 //  *dist = d1 / 100;          //convert to meters
 //  *bearing = d2 /  100;      //convert to degrees
 //}
-
-////////////////////////////////////////////////////////////////////////////////////
-// Calculate our current speed vector from gps position data
-//
-static void GPS_calc_velocity(void)
-{
-    static int16_t speed_old[2] = { 0, 0 };
-    static int32_t last_coord[2] = { 0, 0 };
-    static uint8_t init = 0;
-
-    if (init) {
-        float tmp = 1.0f / dTnav;
-        actual_speed[GPS_X] = (float)(gpsSol.llh.lon - last_coord[LON]) * GPS_scaleLonDown * tmp;
-        actual_speed[GPS_Y] = (float)(gpsSol.llh.lat - last_coord[LAT]) * tmp;
-
-        actual_speed[GPS_X] = (actual_speed[GPS_X] + speed_old[GPS_X]) / 2;
-        actual_speed[GPS_Y] = (actual_speed[GPS_Y] + speed_old[GPS_Y]) / 2;
-
-        speed_old[GPS_X] = actual_speed[GPS_X];
-        speed_old[GPS_Y] = actual_speed[GPS_Y];
-    }
-    init = 1;
-
-    last_coord[LON] = gpsSol.llh.lon;
-    last_coord[LAT] = gpsSol.llh.lat;
-}
 
 ////////////////////////////////////////////////////////////////////////////////////
 // Calculate a location error between two gps coordinates
@@ -536,7 +368,6 @@ static void GPS_calc_location_error(int32_t *target_lat, int32_t *target_lng, in
     error[LAT] = *target_lat - *gps_lat;        // Y Error
 }
 
-#ifdef USE_NAV
 ////////////////////////////////////////////////////////////////////////////////////
 // Calculate nav_lat and nav_lon from the x and y error and the speed
 //
@@ -642,7 +473,6 @@ static uint16_t GPS_calc_desired_speed(uint16_t max_speed, bool _slow)
     }
     return max_speed;
 }
-#endif
 
 ////////////////////////////////////////////////////////////////////////////////////
 // Utilities
@@ -656,7 +486,6 @@ static int32_t wrap_18000(int32_t error)
     return error;
 }
 
-#ifdef USE_NAV
 static int32_t wrap_36000(int32_t angle)
 {
     if (angle > 36000)
@@ -665,7 +494,6 @@ static int32_t wrap_36000(int32_t angle)
         angle += 36000;
     return angle;
 }
-#endif
 
 void updateGpsStateForHomeAndHoldMode(void)
 {
