+++ src/main/flight/navigation.h
@@ -18,13 +18,8 @@
 #pragma once
 
 #include "common/axis.h"
+#include "io/gps.h"
 
-// navigation mode
-typedef enum {
-    NAV_MODE_NONE = 0,
-    NAV_MODE_POSHOLD,
-    NAV_MODE_WP
-} navigationMode_e;
 
 // FIXME ap_mode is badly named, it's a value that is compared to rcCommand, not a flag at it's name implies.
 
@@ -40,22 +35,14 @@ typedef struct navigationConfig_s {
 
 PG_DECLARE(navigationConfig_t, navigationConfig);
 
-extern int16_t GPS_angle[ANGLE_INDEX_COUNT];                // it's the angles that must be applied for GPS correction
 
-extern int32_t GPS_home[2];
 extern int32_t GPS_hold[2];
 
-extern uint16_t GPS_distanceToHome;        // distance to home point in meters
-extern int16_t GPS_directionToHome;        // direction to home or hol point in degrees
-
-extern navigationMode_e nav_mode;          // Navigation mode
 
 void navigationInit(void);
-void GPS_reset_home_position(void);
 void GPS_reset_nav(void);
 void GPS_set_next_wp(int32_t* lat, int32_t* lon);
 void gpsUsePIDs(struct pidProfile_s *pidProfile);
 void updateGpsStateForHomeAndHoldMode(void);
 void updateGpsWaypointsAndMode(void);
 
-void onGpsNewData(void);
